+++ Infra/Repository/EventParticipationRepository.cs
@@ -1,5 +1,4 @@
 using Microsoft.EntityFrameworkCore;
-using MyVocaList.Infra.Utils;
 using MyVocaList.Domain.Entity;
 using MyVocaList.Domain.RepositoryInterface;
 
@@ -14,8 +13,8 @@ namespace MyVocaList.Infra.Repository
 
         public async Task<IEnumerable<EventParticipation>> GetParticipationsByPersonIdAndEventIdAsync(int personId, int eventId)
         {
-            Guard.AgainstNegativeOrZero(personId, nameof(personId));
-            Guard.AgainstNegativeOrZero(eventId, nameof(eventId));
+            ArgumentOutOfRangeException.ThrowIfNegativeOrZero(personId, nameof(personId));
+            ArgumentOutOfRangeException.ThrowIfNegativeOrZero(eventId, nameof(eventId));
 
             return await _dbSet
                 .Where(ep => ep.PersonId == personId && ep.EventId == eventId)